trip duration in date utils
